% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_ratings.R
\name{explore_ratings}
\alias{explore_ratings}
\title{"Explore Ratings"}
\usage{
explore_ratings(appstore, appname, size)
}
\arguments{
\item{appstore}{"A string name for the app e.g safeboda"}

\item{appname}{"A string name for the app store e.g google"}

\item{size}{"A numeric vector of how many rows to return"}
}
\value{
"This function returns a  \code{dataset} of ratings from the appstore"
}
\description{
This is a function named 'explore_ratings' which will load a dataset of of an app you choose
}
\note{
Experimental
}
\examples{
explore_ratings("google", "safeboda", 10)

}
\author{
"My Name is Simon Kazooba"
}
