% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstorr.R
\docType{package}
\name{appstorr}
\alias{appstorr}
\alias{appstorr-package}
\title{appstorr: "A package for analyzing appstore ratings"}
\description{
This packages provides power to developers to explore appstore ratings easily
}
\author{
Simon Kazooba \email{smkabz@gmail.com}
}
