% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_report.R
\name{quick_report}
\alias{quick_report}
\title{"Quick Report"}
\usage{
quick_report(appname)
}
\arguments{
\item{appname}{"A string name for the name e.g safeboda"}
}
\value{
"This function returns a quick report of ratings for an \code{App}"
}
\description{
This is a function named 'quick_report' will run a report on the data
}
\note{
Experimental
}
\examples{
quick_report("safeboda")

}
\author{
"My Name is Simon Kazooba"
}
